refactor: 💡 dirs of contents
