+++ src/features/contents/actions/delete-contents.ts
