+++ src/features/contents/components/add-contents-form.tsx
@@ -4,8 +4,8 @@ import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Label } from "@/components/ui/label";
 import { Textarea } from "@/components/ui/textarea";
-import { addContents } from "@/features/con/actions/add-contents";
-import { contentsAtom } from "@/features/con/stores/contents-atom";
+import { addContents } from "@/features/contents/actions/add-contents";
+import { contentsAtom } from "@/features/contents/stores/contents-atom";
 import { useToast } from "@/hooks/use-toast";
 import { useSetAtom } from "jotai";
 import { ClipboardPasteIcon } from "lucide-react";
