+++ src/features/contents/utils/validate-contents.ts
@@ -1,6 +1,6 @@
 import "server-only";
 import { InvalidFormatError } from "@/error-classes";
-import { contentsSchema } from "@/features/con/schemas/contents-schema";
+import { contentsSchema } from "@/features/contents/schemas/contents-schema";
 
 export function validateContents(formData: FormData) {
 	const contentsValidatedFields = contentsSchema.safeParse({
